Improve polling (#32)
Now we only sleep for the defined interval when we either:

1. Encounter an empty event batch; or
2. Encounter an error getting the batch

This way, we don't unnecessarily sleep between batches of events when there are still events to process. The only trade-off here is that we would perform 1 additional query after the last batch in order to determine whether there are more events to process, but this is not a concern. In future, we may improve the sleep handling to back off for a longer period when multiple poll operations return no events, but it can get messy.